Move hasFinishedLoadingGTMContainers to Analytics datastore.
