+++ assets/js/modules/analytics/datastore/setup-flow.test.js
@@ -34,7 +34,12 @@ import {
 	unsubscribeFromAll,
 	provideModules,
 	untilResolved,
+	freezeFetch,
 } from '../../../../../tests/js/utils';
+import { CORE_MODULES } from '../../../googlesitekit/modules/datastore/constants';
+import { MODULES_TAGMANAGER } from '../../tagmanager/datastore/constants';
+import { parseLiveContainerVersionIDs } from '../../tagmanager/datastore/__factories__/utils';
+import * as gtmFixtures from '../../tagmanager/datastore/__fixtures__';
 
 const accountID = 'pub-12345678';
 
@@ -332,6 +337,186 @@ describe( 'modules/analytics setup-flow', () => {
 					.canUseGA4Controls();
 
 				expect( canUseControls ).toBe( true );
+			} );
+		} );
+
+		describe( 'hasFinishedLoadingGTMContainers', () => {
+			it( 'should return TRUE when the GTM module is not connected', () => {
+				// Set tagmanager module to not connected.
+				registry.dispatch( CORE_MODULES ).receiveGetModules( [
+					{
+						slug: 'analytics',
+						name: 'Analytics',
+						active: true,
+					},
+					{
+						slug: 'tagmanager',
+						name: 'Tag Manager',
+						active: false,
+						connected: false,
+					},
+				] );
+
+				const hasFinishedLoading = registry
+					.select( MODULES_ANALYTICS )
+					.hasFinishedLoadingGTMContainers();
+
+				expect( hasFinishedLoading ).toBe( true );
+			} );
+
+			it( 'should return TRUE when the GTM module is connected and the web containers are loaded', async () => {
+				registry.dispatch( CORE_MODULES ).receiveGetModules( [
+					{
+						slug: 'analytics',
+						name: 'Analytics',
+						active: true,
+					},
+					{
+						slug: 'tagmanager',
+						name: 'Tag Manager',
+						active: true,
+						connected: true,
+					},
+				] );
+
+				const webContainerVersion =
+					gtmFixtures.liveContainerVersions.web.gaWithVariable;
+				const gtmAccountID = webContainerVersion.accountId; // eslint-disable-line sitekit/acronym-case
+				const gtmContainerID = webContainerVersion.containerId; // eslint-disable-line sitekit/acronym-case
+
+				parseLiveContainerVersionIDs(
+					webContainerVersion,
+					( { internalContainerID, containerID } ) => {
+						registry.dispatch( MODULES_TAGMANAGER ).setSettings( {
+							accountID: gtmAccountID,
+							containerID,
+							internalContainerID,
+							useSnippet: true,
+						} );
+						registry
+							.dispatch( MODULES_TAGMANAGER )
+							.receiveGetLiveContainerVersion(
+								webContainerVersion,
+								{
+									accountID: gtmAccountID,
+									internalContainerID,
+								}
+							);
+						registry
+							.select( MODULES_TAGMANAGER )
+							.getLiveContainerVersion(
+								gtmAccountID,
+								internalContainerID
+							);
+					}
+				);
+
+				await untilResolved(
+					registry,
+					MODULES_TAGMANAGER
+				).getLiveContainerVersion( gtmAccountID, gtmContainerID );
+				const hasFinishedLoading = registry
+					.select( MODULES_ANALYTICS )
+					.hasFinishedLoadingGTMContainers();
+
+				expect( hasFinishedLoading ).toBe( true );
+			} );
+
+			it( 'should return TRUE when the GTM module is available and the AMP containers are loaded', async () => {
+				registry.dispatch( CORE_MODULES ).receiveGetModules( [
+					{
+						slug: 'analytics',
+						name: 'Analytics',
+						active: true,
+					},
+					{
+						slug: 'tagmanager',
+						name: 'Tag Manager',
+						active: true,
+						connected: true,
+					},
+				] );
+
+				const ampContainerVersion =
+					gtmFixtures.liveContainerVersions.amp.ga;
+				const gtmAccountID = ampContainerVersion.accountId; // eslint-disable-line sitekit/acronym-case
+				const gtmAMPContainerID = ampContainerVersion.containerId; // eslint-disable-line sitekit/acronym-case
+
+				parseLiveContainerVersionIDs(
+					ampContainerVersion,
+					( { internalAMPContainerID, containerID } ) => {
+						registry.dispatch( MODULES_TAGMANAGER ).setSettings( {
+							accountID: gtmAccountID,
+							containerID,
+							internalContainerID: internalAMPContainerID,
+							useSnippet: true,
+						} );
+						registry
+							.dispatch( MODULES_TAGMANAGER )
+							.setInternalAMPContainerID(
+								internalAMPContainerID
+							);
+						registry
+							.dispatch( MODULES_TAGMANAGER )
+							.receiveGetLiveContainerVersion(
+								ampContainerVersion,
+								{
+									accountID: gtmAccountID,
+									internalContainerID: internalAMPContainerID,
+								}
+							);
+						registry
+							.select( MODULES_TAGMANAGER )
+							.getLiveContainerVersion(
+								gtmAccountID,
+								internalAMPContainerID
+							);
+					}
+				);
+
+				await untilResolved(
+					registry,
+					MODULES_TAGMANAGER
+				).getLiveContainerVersion( gtmAccountID, gtmAMPContainerID );
+				const hasFinishedLoading = registry
+					.select( MODULES_ANALYTICS )
+					.hasFinishedLoadingGTMContainers();
+
+				expect( hasFinishedLoading ).toBe( true );
+			} );
+
+			it( 'should return FALSE when the selector is not resolved yet', () => {
+				registry.dispatch( CORE_MODULES ).receiveGetModules( [
+					{
+						slug: 'analytics',
+						name: 'Analytics',
+						active: true,
+					},
+					{
+						slug: 'tagmanager',
+						name: 'Tag Manager',
+						active: true,
+						connected: true,
+					},
+				] );
+				registry.dispatch( MODULES_TAGMANAGER ).setSettings( {
+					accountID: 100,
+					containerID: 200,
+					internalContainerID: 200,
+					useSnippet: true,
+				} );
+
+				freezeFetch(
+					new RegExp(
+						'^/google-site-kit/v1/modules/tagmanager/data/live-container-version'
+					)
+				);
+
+				const hasFinishedLoading = registry
+					.select( MODULES_ANALYTICS )
+					.hasFinishedLoadingGTMContainers();
+
+				expect( hasFinishedLoading ).toBe( false );
 			} );
 		} );
 	} );
